feat(voice): in-modal views + bring-your-own-key path on the voice first-run card (JARVIS-1288)
Adds an "I have my own STT/TTS API key" affordance to the voice-mode
first-use card, and restructures the card so its detours are views of the
one modal rather than modals stacked on top of it.

- The card now has three views: intro, voice picker, and bring-your-own-key.
  Selecting a voice no longer opens a second dialog over the card; both
  detours swap the card's header/body/footer and offer a back arrow. Width is
  held constant so navigating doesn't resize the dialog.
- The BYOK link is a quiet text link in the footer, left of "Start talking" —
  managed speech stays the obvious path.
- Saving a key returns to the intro view, so the user lands back on
  "Start talking".
- Escape inside a sub-view navigates back instead of cancelling, so a stray
  keypress can't discard a half-entered API key. Dismissing from the intro
  still leaves `firstRunSeen` un-consumed.
- The back arrow is in-modal navigation, not a cancel, so it stays available
  under the Capacitor iOS `nonDismissible` lock.

The provider/key forms are the settings page's own, extracted from
SpeechToTextCard/TextToSpeechCard so there is one implementation. They move
to top-level `components/speech/` because two domains render them now and
domain peers must not import each other (CONVENTIONS.md § Top-level shared
directories); the settings cards become thin ByoServiceCard wrappers. The
`domains/settings/ai/` helpers those forms still use are shared by the whole
settings AI page and are left in place.

The voice-room settings popover keeps VoicePickerModal — it opens from a
popover, not a modal, so there is no stacking there.
