+++ clients/web/src/domains/chat/voice/voice-room/voice-first-run-card.test.tsx
@@ -1,9 +1,10 @@
 /**
  * Tests for `VoiceFirstRunCard`.
  *
- * The card is exercised in isolation: `onStart` is a spy. The assistant-avatar
- * hook is stubbed so the card renders without the React Query graph — the
- * avatar is chrome, not behavior.
+ * The card is exercised in isolation: `onStart` is a spy, and the assistant
+ * avatar / managed-voice / BYOK-form dependencies are stubbed so the card
+ * renders without the React Query graph — they are chrome around the card's own
+ * behavior, and each has its own tests.
  *
  * Load-bearing behavior:
  *   - the card renders on first run and does NOT start on its own,
@@ -11,7 +12,9 @@
  *     Settings, not front-loaded here,
  *   - "Start" invokes the caller's `onStart`; wiring that `onStart` to
  *     `markFirstRunSeen` (as the composer does) consumes the first run so a
- *     second entry would skip the card.
+ *     second entry would skip the card,
+ *   - the voice and bring-your-own-key detours are VIEWS of this one modal, not
+ *     modals stacked on it, and both return to the intro.
  */
 import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 
@@ -28,24 +31,57 @@ mock.module("@/hooks/use-assistant-avatar", () => ({
     invalidate: () => {},
   }),
 }));
-// The voice picker's own tests cover it; here it stays collapsed (unavailable)
-// so the card renders without the daemon query graph / a QueryClient.
-mock.module("@/domains/chat/voice/voice-room/use-managed-voice-selection", () => ({
-  useManagedVoiceSelection: () => ({
-    available: false,
-    voices: [],
-    currentModel: "",
-    selectModel: () => {},
-    selecting: false,
+
+// Managed voice availability drives whether the Voice row (and so the voice
+// view) exists at all. Mutable so a test can turn it on without a second
+// module mock — `VoiceSettingRow` and `VoiceList` both read this hook.
+let managedVoiceAvailable = false;
+const selectModel = mock((_model: string) => {});
+mock.module(
+  "@/domains/chat/voice/voice-room/use-managed-voice-selection",
+  () => ({
+    useManagedVoiceSelection: () => ({
+      available: managedVoiceAvailable,
+      voices: managedVoiceAvailable
+        ? [
+            // `<accent> · <traits>` — the order splitVoiceDescription expects.
+            { model: "aura-2-thalia-en", description: "American · warm, clear", label: "Thalia", source: "deepgram", sampleUrl: "" },
+            { model: "aura-2-orion-en", description: "American · calm, low", label: "Orion", source: "deepgram", sampleUrl: "" },
+          ]
+        : [],
+      currentModel: managedVoiceAvailable ? "aura-2-thalia-en" : "",
+      selectModel,
+      selecting: false,
+    }),
   }),
+);
+
+// The BYOK forms own the daemon config/credential graph and are covered by the
+// settings-page tests; here they stand in as a save affordance so the card's
+// navigation is what's under test.
+mock.module("@/components/speech/stt-provider-form", () => ({
+  SttProviderForm: ({ onSaved }: { onSaved?: () => void }) => (
+    <button type="button" onClick={onSaved}>
+      Save STT
+    </button>
+  ),
+}));
+mock.module("@/components/speech/tts-provider-form", () => ({
+  TtsProviderForm: ({ onSaved }: { onSaved?: () => void }) => (
+    <button type="button" onClick={onSaved}>
+      Save TTS
+    </button>
+  ),
 }));
 
 import { useVoicePrefsStore } from "@/stores/voice-prefs-store";
 
-// Imported after the mock so the card resolves against the stubbed hook.
+// Imported after the mocks so the card resolves against the stubs.
 const { VoiceFirstRunCard } = await import(
   "@/domains/chat/voice/voice-room/voice-first-run-card"
 );
+
+const BYOK_LINK = "I have my own STT/TTS API key";
 
 afterEach(cleanup);
 beforeEach(() => {
@@ -55,6 +91,8 @@ beforeEach(() => {
     showAssistantTranscript: false,
     firstRunSeen: false,
   });
+  managedVoiceAvailable = false;
+  selectModel.mockClear();
 });
 
 describe("VoiceFirstRunCard", () => {
@@ -125,5 +163,91 @@ describe("VoiceFirstRunCard", () => {
     );
     expect(queryByLabelText("Close")).toBeNull();
     expect(getByText("Start talking")).toBeTruthy();
+  });
+
+  describe("bring-your-own-key view", () => {
+    test("the link opens the key view in place — one dialog, no stack", () => {
+      const { getByText, queryByText, baseElement } = render(
+        <VoiceFirstRunCard assistantId="asst_test" onStart={() => {}} />,
+      );
+
+      fireEvent.click(getByText(BYOK_LINK));
+
+      expect(getByText("Use your own API keys")).toBeTruthy();
+      expect(getByText("Speech to text")).toBeTruthy();
+      expect(getByText("Text to speech")).toBeTruthy();
+      // The intro's forward action is gone — this is a view swap, not an
+      // overlay on top of the intro.
+      expect(queryByText("Start talking")).toBeNull();
+      expect(baseElement.querySelectorAll('[role="dialog"]').length).toBe(1);
+    });
+
+    test("back returns to the intro without starting or consuming the first run", () => {
+      const onStart = mock(() => {});
+      const { getByText, getByLabelText } = render(
+        <VoiceFirstRunCard assistantId="asst_test" onStart={onStart} />,
+      );
+
+      fireEvent.click(getByText(BYOK_LINK));
+      fireEvent.click(getByLabelText("Back"));
+
+      expect(getByText("Start talking")).toBeTruthy();
+      expect(onStart).not.toHaveBeenCalled();
+      expect(useVoicePrefsStore.getState().firstRunSeen).toBe(false);
+    });
+
+    test("saving a key lands back on the Start talking view", () => {
+      const { getByText } = render(
+        <VoiceFirstRunCard assistantId="asst_test" onStart={() => {}} />,
+      );
+
+      fireEvent.click(getByText(BYOK_LINK));
+      fireEvent.click(getByText("Save STT"));
+
+      expect(getByText("Start talking")).toBeTruthy();
+    });
+
+    test("the link is reachable under the iOS lock too", () => {
+      // The lock removes cancels, not in-modal navigation — a locked card
+      // still has to let a BYOK user configure before the mic prompt.
+      const { getByText, getByLabelText } = render(
+        <VoiceFirstRunCard
+          assistantId="asst_test"
+          onStart={() => {}}
+          nonDismissible
+        />,
+      );
+
+      fireEvent.click(getByText(BYOK_LINK));
+      expect(getByText("Use your own API keys")).toBeTruthy();
+      fireEvent.click(getByLabelText("Back"));
+      expect(getByText("Start talking")).toBeTruthy();
+    });
+  });
+
+  describe("voice view", () => {
+    test("the Voice row opens the picker in place, and choosing returns to the intro", () => {
+      managedVoiceAvailable = true;
+      const { getByText, getByLabelText, baseElement } = render(
+        <VoiceFirstRunCard assistantId="asst_test" onStart={() => {}} />,
+      );
+
+      fireEvent.click(getByText("Voice"));
+
+      // Same single dialog, swapped to the picker view.
+      expect(baseElement.querySelectorAll('[role="dialog"]').length).toBe(1);
+      expect(getByLabelText("Assistant voice")).toBeTruthy();
+
+      fireEvent.click(getByText("Calm, low"));
+      expect(selectModel).toHaveBeenCalledWith("aura-2-orion-en");
+      expect(getByText("Start talking")).toBeTruthy();
+    });
+
+    test("collapses entirely when managed voice selection is unavailable", () => {
+      const { queryByText } = render(
+        <VoiceFirstRunCard assistantId="asst_test" onStart={() => {}} />,
+      );
+      expect(queryByText("Voice")).toBeNull();
+    });
   });
 });
