+++ clients/web/src/domains/chat/voice/voice-room/voice-setting-row.tsx
@@ -1,11 +1,12 @@
 /**
  * The "Voice" row shown in the voice-room settings popover and the first-run
- * card: the assistant's current voice with a chevron, opening the dedicated
- * {@link VoicePickerModal} (the full catalog with previews doesn't fit inline).
+ * card: the assistant's current voice with a chevron, leading to the full
+ * catalog with previews (which doesn't fit inline).
  *
- * The parent owns the modal — this row only reports the click via `onOpen` — so
- * closing the settings popover (which unmounts this row) can't unmount the modal
- * with it.
+ * The parent owns where that leads — this row only reports the click via
+ * `onOpen`. The settings popover opens {@link VoicePickerModal} (so closing the
+ * popover, which unmounts this row, can't unmount the picker with it); the
+ * first-run card swaps to its own voice view instead.
  *
  * Renders nothing unless managed voice selection is available (managed assistant
  * + a daemon that offers it); BYO providers choose their voice in Settings.
